consumer/pyproject.toml
[project]
name = "automata-consumer"
version = "0.1.0"
description = "AutomataCodex Consumer サービス"
requires-python = ">=3.11"
dependencies = [
    "automata-shared",
    "aio-pika>=9.0",
    "python-gitlab>=4.0",
    "docker>=7.0",
    "mcp>=1.0",
    "tiktoken>=0.7",
    "agent-framework-core>=1.0.0rc1",
]

[tool.uv.sources]
automata-shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
